requirements.txt
streamlit
pandas
scikit-learn
matplotlib
wordcloud
openai
joblib
nltk
google-generativeai